Crud 100% de PESSOA/FUNCIONARIO/PROFESSOR/MDALIDADE

+++ src/main/java/academia/model/Modalidade.java
@@ -21,7 +21,7 @@ public class Modalidade {
 	public Modalidade() {
 		super();
 	}
-	public long getIdModalidade() {
+	public int getIdModalidade() {
 		return idModalidade;
 	}
 	public void setIdModalidade(int idModalidade) {
